requirements.txt
matplotlib==3.7.1
mysql-connector-python==8.0.33
python-telegram-bot==20.3
